requirements_face_api.txt
flask
flask-cors
torch
torchvision
ultralytics
opencv-python
pillow
numpy
